// apps/admin-dashboard/src/types/enums.ts
export enum UserRole {
  Customer = 0,
  Admin = 1,
  Dev = 2,
}

export enum FileType {
  Document = 0,
  Image = 1,
  Video = 2,
  Audio = 3,
  Archive = 4,
  Other = 5,
}

export enum FolderType {
  General = 0,
  Images = 1,
  Documents = 2,
  Videos = 3,
  Audio = 4,
  UserAvatars = 5,
  CompanyAssets = 6,
  Temporary = 7,
}